fix: prevent throwing on duplicate txIds during descriptor update in discovery
This fix removes the error thrown when a `txId` has already been pushed or is part of the discovery object. Instead of throwing, it now checks whether the `txId` is already present in the `outputData.txIds` array and only pushes the `txId` if it's not already included. This prevents from throwing when auto-sending to owned outputs.